Code review fix: Moved error prototype related code to its own file
