+++ lib/err-with-cause.js
@@ -2,7 +2,8 @@
 
 module.exports = errWithCauseSerializer
 
-const { isErrorLike, pinoErrProto, pinoErrorSymbols } = require('./err-helpers')
+const { isErrorLike } = require('./err-helpers')
+const { pinoErrProto, pinoErrorSymbols } = require('./err-proto')
 const { seen } = pinoErrorSymbols
 
 const { toString } = Object.prototype
