+++ lib/err.js
@@ -2,7 +2,8 @@
 
 module.exports = errSerializer
 
-const { messageWithCauses, stackWithCauses, isErrorLike, pinoErrProto, pinoErrorSymbols } = require('./err-helpers')
+const { messageWithCauses, stackWithCauses, isErrorLike } = require('./err-helpers')
+const { pinoErrProto, pinoErrorSymbols } = require('./err-proto')
 const { seen } = pinoErrorSymbols
 
 const { toString } = Object.prototype
